+++ services/history/package.json
@@ -1,0 +1,47 @@
+{
+  "name": "history",
+  "version": "0.0.0",
+  "main": "index.js",
+  "scripts": {
+    "build": "npx tsc",
+    "dev": "node src/index.ts",
+    "start": "nodemon src/index.ts",
+    "lint": "npx eslint .",
+    "lint:fix": "npx eslint . --fix",
+    "test": "mocha --require ts-node/register tests/**/*.spec.ts --exit"
+  },
+  "author": "",
+  "license": "ISC",
+  "description": "",
+  "dependencies": {
+    "body-parser": "^1.20.3",
+    "cors": "^2.8.5",
+    "express": "^4.21.0",
+    "mongoose": "^8.6.2",
+    "morgan": "^1.10.0"
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.10.0",
+    "@types/chai": "^4.3.19",
+    "@types/cors": "^2.8.17",
+    "@types/eslint__js": "^8.42.3",
+    "@types/express": "^4.17.21",
+    "@types/mocha": "^10.0.8",
+    "@types/morgan": "^1.9.9",
+    "@types/node": "^22.5.4",
+    "@typescript-eslint/eslint-plugin": "^8.5.0",
+    "chai": "^4.5.0",
+    "chai-http": "^4.4.0",
+    "eslint": "^9.10.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "mocha": "^10.7.3",
+    "mongodb-memory-server": "^10.0.0",
+    "nodemon": "^3.1.4",
+    "prettier": "3.3.3",
+    "prettier-eslint": "^16.3.0",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.5.4",
+    "typescript-eslint": "^8.5.0"
+  }
+}
