+++ services/user-service/index.js
@@ -1,8 +1,8 @@
 import express from "express";
 import cors from "cors";
 
-import userRoutes from "./routes/user-routes.js";
-import authRoutes from "./routes/auth-routes.js";
+import userRoutes from "./src/routes/user-routes.js";
+import authRoutes from "./src/routes/auth-routes.js";
 
 const app = express();
