--- services/user-service/model/repository.js
@@ -1,71 +1,0 @@
-import UserModel from "./user-model.js";
-import "dotenv/config";
-import { connect } from "mongoose";
-
-export async function connectToDB() {
-  let mongoDBUri =
-    process.env.ENV === "PROD"
-      ? process.env.USER_SERVICE_CLOUD_URI
-      : process.env.USER_SERVICE_LOCAL_URI;
-
-  await connect(mongoDBUri);
-}
-
-export async function createUser(username, email, password) {
-  return new UserModel({ username, email, password }).save();
-}
-
-export async function findUserByEmail(email) {
-  return UserModel.findOne({ email });
-}
-
-export async function findUserById(userId) {
-  return UserModel.findById(userId);
-}
-
-export async function findUserByUsername(username) {
-  return UserModel.findOne({ username });
-}
-
-export async function findUserByUsernameOrEmail(username, email) {
-  return UserModel.findOne({
-    $or: [
-      { username },
-      { email },
-    ],
-  });
-}
-
-export async function findAllUsers() {
-  return UserModel.find();
-}
-
-export async function updateUserById(userId, username, email, password) {
-  return UserModel.findByIdAndUpdate(
-    userId,
-    {
-      $set: {
-        username,
-        email,
-        password,
-      },
-    },
-    { new: true },  // return the updated user
-  );
-}
-
-export async function updateUserPrivilegeById(userId, isAdmin) {
-  return UserModel.findByIdAndUpdate(
-    userId,
-    {
-      $set: {
-        isAdmin,
-      },
-    },
-    { new: true },  // return the updated user
-  );
-}
-
-export async function deleteUserById(userId) {
-  return UserModel.findByIdAndDelete(userId);
-}
