--- services/user-service/model/user-model.js
@@ -1,31 +1,0 @@
-import mongoose from "mongoose";
-
-const Schema = mongoose.Schema;
-
-const UserModelSchema = new Schema({
-  username: {
-    type: String,
-    required: true,
-    unique: true,
-  },
-  email: {
-    type: String,
-    required: true,
-    unique: true,
-  },
-  password: {
-    type: String,
-    required: true,
-  },
-  createdAt: {
-    type: Date,
-    default: Date.now, // Setting default to the current date/time
-  },
-  isAdmin: {
-    type: Boolean,
-    required: true,
-    default: false,
-  },
-});
-
-export default mongoose.model("UserModel", UserModelSchema);
