--- services/user-service/routes/user-routes.js
@@ -1,27 +1,0 @@
-import express from "express";
-
-import {
-  createUser,
-  deleteUser,
-  getAllUsers,
-  getUser,
-  updateUser,
-  updateUserPrivilege,
-} from "../controller/user-controller.js";
-import { verifyAccessToken, verifyIsAdmin, verifyIsOwnerOrAdmin } from "../middleware/basic-access-control.js";
-
-const router = express.Router();
-
-router.get("/", verifyAccessToken, verifyIsAdmin, getAllUsers);
-
-router.patch("/:id/privilege", verifyAccessToken, verifyIsAdmin, updateUserPrivilege);
-
-router.post("/", createUser);
-
-router.get("/:id", verifyAccessToken, verifyIsOwnerOrAdmin, getUser);
-
-router.patch("/:id", verifyAccessToken, verifyIsOwnerOrAdmin, updateUser);
-
-router.delete("/:id", verifyAccessToken, verifyIsOwnerOrAdmin, deleteUser);
-
-export default router;
