+++ services/user-service/src/GuideAssets/Creation.png
