+++ services/user-service/src/GuideAssets/SidePane.png
