+++ services/user-service/src/routes/auth-routes.js
@@ -1,0 +1,12 @@
+import express from "express";
+
+import { handleLogin, handleVerifyToken } from "../controller/auth-controller.js";
+import { verifyAccessToken } from "../middleware/basic-access-control.js";
+
+const router = express.Router();
+
+router.post("/login", handleLogin);
+
+router.get("/verify-token", verifyAccessToken, handleVerifyToken);
+
+export default router;
